requirements.txt
anyio==3.6.2
apt-xapian-index==0.49
blinker==1.5
Brlapi==0.8.4
catfish==4.16.4
certifi==2022.9.24
chardet==5.1.0
click==8.1.3
colorama==0.4.6
cryptography==38.0.4
cupshelpers==1.0
dbus-python==1.3.2
deprecation==2.0.7
distro==1.8.0
distro-info==1.5+deb12u1
dnspython==2.3.0
dtrx==8.4.0
ewmh==0.1.6
gattlib==0.20210616
h11==0.14.0
httpcore==0.16.3
httplib2==0.20.4
httpx==0.23.3
idna==3.3
img2pdf==0.4.4
lazr.restfulclient==0.14.5
lazr.uri==1.0.6
lightdm-gtk-greeter-settings==1.2.2
louis==3.24.0
lxml==4.9.2
Magnus==1.0.3
markdown-it-py==2.1.0
mdurl==0.1.2
menulibre==2.2.2
mugshot==0.4.3
nala==0.14.0
notify2==0.3
oauthlib==3.2.2
onboard==1.4.1
packaging==23.0
pdfarranger==1.10.0
pexpect==4.8.0
pikepdf==6.0.0+dfsg
Pillow==9.4.0
psutil==5.9.4
ptyprocess==0.7.0
PyBluez==0.23
pycairo==1.20.1
pycups==2.0.1
Pygments==2.14.0
PyGObject==3.42.2
PyJWT==2.6.0
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-sip==12.11.1
pysmbc==1.0.23
python-apt==2.6.0
python-dateutil==2.8.2
python-debian==0.1.49
python-xlib==0.33
pyxdg==0.28
reportlab==3.6.12
rfc3986==1.5.0
rich==13.3.1
rpl==1.14
setproctitle==1.3.1
six==1.16.0
sniffio==1.2.0
socksio==1.0.0
tomli==2.0.1
typer==0.7.0
typing_extensions==4.4.0
ufw==0.36.2
unattended-upgrades==0.1
urllib3==1.26.12
wadllib==1.3.6
xdg==5
